Everything runs ok till the end, ensemble missing
