+++ To-Do
@@ -1,9 +1,7 @@
 07.12.2023
-# 1. Normalize data after holdout
-# 2. Fix SVM hyperparameters
-# 3. Ensemble
-# 4. Save models
-# 5. Hyperparameter Plots
-# 6. Use SMOTE algorithm instead of undersampling?
-# 7. Sum up all libraries used in first cell?
-# 8. Stratified K-Fold Cross Validation for unbalanced data
+# 1. Implement zero-mean normalization
+# 2. Save models
+# 3. Hyperparameter Plots
+# 4. Use SMOTE algorithm instead of undersampling?
+# 5. Sum up all libraries used in first cell?
+# 6. Stratified K-Fold Cross Validation for unbalanced data
